python/client/pyproject.toml
[tool.poetry]
name = "conference-go-client"
version = "0.1.0"
description = "client for conference.go (a conference server writen by golang)"
authors = ["vipcxj <vipcxj@126.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiortc = "^1.5.0"
python-socketio = {extras = ["asyncio-client"], version = "^5.10.0"}


[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"
priority = "default"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
